backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
numpy>=1.24.3
python-dotenv>=1.0.0
langchain>=0.0.312
langchain-openai>=0.0.2
pydantic>=2.4.2
typing-extensions>=4.7.1
wikipedia>=1.4.0
duckduckgo-search>=3.9.9
